progress
updates counter very much unresponsive, voice and event counters somewhat accurate

+++ src/menu.c
@@ -275,7 +275,6 @@ void pauseSong(void) {
             paused = 1;
         } else {
             alCSPPlay(SeqPlayers[slot]);
-            // resetMetrics();
             preventSongPlaying = 1;
             paused = 0;
         }
@@ -501,7 +500,7 @@ Gfx* displayMusicMenu(Gfx* dl) {
     dl = renderMenuOption(dl, volume_eq_strs[volume_equalization], MENUOP_EQ, 20, CONTROLS_Y + 80);
     dl = renderMenuOption(dl, zip_states[zip_in_progress], MENUOP_ZIP, 170, CONTROLS_Y + 80);
     
-    dk_strFormat(voices_in_use_str, "VOICES: %s OF 44", padNumber((int) getVoicesUsed(), 2));
+    dk_strFormat(voices_in_use_str, "VOICES: %s OF 44", padNumber(getVoicesUsed(), 2));
     dk_strFormat(events_in_queue_str, "EVENT QUEUE: %s OF 64 WIP", padNumber(getEventsUsed(), 2));
     dk_strFormat(updates_in_use_str, "UPDATES: %s OF 112", padNumber(getUpdatesUsed(), 3));
